The main purpose of this CL is to get rid of @SuppressWarning with unchecked casting where it is possible. Mostly, it is problems with cloning.
Two util classes were introduced: ObjectUtil for cloning objects and CollectionUtil for making lists deep copy. I change all cloneable CAPARF classes to implement BaseCloneable. This is required to avoid performance problems. Java Cloneable interface does not make clone() method visible (it is protected method of Object) and it is required to use reflection in order to clone objects that implements Cloneable. It is 2x slower compared to direct call.